Move chdirToProjectRoot to load-config
The helper derives the project root from the config file, it has nothing to do
with sockets. With it living next to loadConfig the socket file can simply be
referenced by its plain name again, so the socket util file is gone.
