+++ src/commands/auto-start-daemon.ts
@@ -2,8 +2,7 @@ import { spawn } from "child_process";
 import { existsSync, unlinkSync } from "fs";
 import { createConnection } from "net";
 
-import { loadConfig } from "../utils/load-config.js";
-import { chdirToProjectRoot, SOCKET_FILE_NAME } from "../utils/socket.js";
+import { chdirToProjectRoot, loadConfig } from "../utils/load-config.js";
 
 function isDaemonRunning(): Promise<boolean> {
     return new Promise((resolve, reject) => {
@@ -11,7 +10,7 @@ function isDaemonRunning(): Promise<boolean> {
             client.destroy();
             resolve(false);
         }, 5000);
-        const client = createConnection(SOCKET_FILE_NAME);
+        const client = createConnection(`.pm.sock`);
         client.on("connect", () => {
             clearTimeout(timeout);
             client.destroy();
@@ -31,18 +30,18 @@ function isDaemonRunning(): Promise<boolean> {
 export async function autoStartDaemon(): Promise<void> {
     const { sources } = await loadConfig();
     chdirToProjectRoot(sources[0]);
-    if (existsSync(SOCKET_FILE_NAME)) {
+    if (existsSync(`.pm.sock`)) {
         if (await isDaemonRunning()) {
             // daemon is running
             return;
         }
         // socket file exists but daemon is not running, remove stale socket file
-        unlinkSync(SOCKET_FILE_NAME);
+        unlinkSync(`.pm.sock`);
     }
     console.log("starting dev-pm daemon...");
     const child = spawn(process.argv[0], [process.argv[1], "start-daemon"], { detached: true, stdio: ["ignore", "ignore", "ignore"] });
     child.unref();
-    while (!existsSync(SOCKET_FILE_NAME)) {
+    while (!existsSync(`.pm.sock`)) {
         await new Promise((r) => setTimeout(r, 100));
     }
 }
