+++ src/commands/connect.ts
@@ -1,7 +1,6 @@
 import { createConnection, type Socket } from "net";
 
-import { loadConfig } from "../utils/load-config.js";
-import { chdirToProjectRoot, SOCKET_FILE_NAME } from "../utils/socket.js";
+import { chdirToProjectRoot, loadConfig } from "../utils/load-config.js";
 import { autoStartDaemon } from "./auto-start-daemon.js";
 
 export async function connect(): Promise<Socket> {
@@ -10,10 +9,10 @@ export async function connect(): Promise<Socket> {
     chdirToProjectRoot(sources[0]);
 
     return new Promise((resolve, reject) => {
-        const client = createConnection(SOCKET_FILE_NAME);
+        const client = createConnection(`.pm.sock`);
         client.on("error", (error: { code: string }) => {
             if (error.code == "ECONNREFUSED") {
-                console.log(`Error connecting to dev-pm daemon at ${SOCKET_FILE_NAME}.`);
+                console.log("Error connecting to dev-pm daemon at .pm.sock.");
                 process.exit(-1);
             }
             reject(error);
