+++ src/utils/load-config.ts
@@ -1,3 +1,4 @@
+import { dirname } from "path";
 import { loadConfig as unconfigLoadConfig } from "unconfig";
 
 import type { Config } from "../config.js";
@@ -22,4 +23,11 @@ export async function loadConfig() {
         throw new Error("dev-pm.config doesn't include required scripts");
     }
     return config;
+}
+
+// The project root is the directory containing the config file. All commands change into it, which also keeps the
+// path of the .pm.sock socket file short: unix socket paths must fit into sockaddr_un.sun_path (108 bytes on Linux,
+// 104 bytes on macOS), so an absolute path in a deeply nested project fails with EINVAL.
+export function chdirToProjectRoot(configSource: string): void {
+    process.chdir(dirname(configSource));
 }
